pyproject.toml
[project]
name = "roof"
version = "0.1.0"
description = "API service for ROOF Pro"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "anyio==4.4.0",
    "aioboto3==14.3.0",
    "asyncpg==0.29.0",
    "aiosmtplib==3.0.2",
    "alembic==1.15.2",
    "fastapi==0.112.0",
    "fastapi-cache2==0.2.2",
    "fastapi-cli==0.0.5",
    "loguru==0.7.2",
    "passlib==1.7.4",
    "prometheus-fastapi-instrumentator==7.0.0",
    "prometheus-client==0.20.0",
    "pydantic==2.8.2",
    "pydantic-extra-types==2.9.0",
    "pydantic-settings==2.4.0",
    "pydantic-core==2.20.1",
    "pydantic[email]==2.8.2",
    "Pygments==2.18.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "python-multipart==0.0.9",
    "pyyaml==6.0.2",
    "redis==4.6.0",
    "SQLAlchemy==2.0.32",
    "starlette==0.37.2",
    "time-machine==2.15.0",
    "uvicorn==0.30.5",
    "websockets==12.0",
    "yarl==1.9.4",
    "shapely==2.0.6",
    "matplotlib==3.9.2",
    "openpyxl==3.1.5",
    "user-agents==2.2.0",
    "unidecode==1.4.0",
    "geoip2==5.1.0",
    "itsdangerous==2.2.0",
    "xlsxwriter==3.2.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.21.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"